refactor: migrate react component types to module
