+++ packages/react/src/component.ts
@@ -1,5 +1,5 @@
 import { event, METHOD, watch } from '@expressive/mvc';
-import { ReactNode } from 'react';
+import React, { FunctionComponent, ReactNode } from 'react';
 
 import { Pragma, ReactState as State } from './state';
 import { provide, Layers } from './context';
@@ -7,7 +7,62 @@ import { Context } from '.';
 
 const OUTER = new WeakMap<Component, Context>();
 
-export class Component extends State {
+type HasProps<T extends State> = {
+  [K in Exclude<keyof T, keyof State>]?: T[K];
+};
+
+type ComponentProps<T extends State> = HasProps<T> & {
+  /**
+   * Callback for newly created instance. Only called once.
+   * @returns Callback to run when instance is destroyed.
+   */
+  is?: (instance: T) => void;
+
+  /**
+   * A fallback react tree to show when suspended.
+   * If not provided, `fallback` property of the State will be used.
+   */
+  fallback?: React.ReactNode;
+};
+
+type Props<T extends State> = T extends {
+  render(props: infer P, self: any): any;
+}
+  ? ComponentProps<T> & Omit<P, keyof AsComponent>
+  : ComponentProps<T> & { children?: React.ReactNode };
+
+/**
+ * Props which will not conflict with a State's use as a Component.
+ *
+ * Built-in properties must be optional, as they will always be omitted.
+ */
+type RenderProps<T extends State> = HasProps<T> & {
+  is?: never;
+  get?: never;
+  set?: never;
+};
+
+/** State which is not incompatable as Component in React. */
+interface AsComponent extends State {
+  render?(props: RenderProps<this>, self: this): React.ReactNode;
+  fallback?: React.ReactNode;
+}
+
+interface FC<
+  T extends State,
+  P extends State.Assign<T>
+> extends FunctionComponent<P & Props<T>> {
+  displayName?: string;
+  State: State.Extends<T>;
+}
+
+declare module './state' {
+  namespace ReactState {
+    export { ComponentProps, Props, RenderProps, AsComponent, FC };
+  }
+}
+
+class Component extends State {
   static contextType = Layers;
 
   private _props!: State.ComponentProps<this>;
@@ -101,3 +156,5 @@ function Render<T extends Component>(this: T, render: () => ReactNode) {
 
   return state[0]();
 }
+
+export { Component, FC, AsComponent, Props, RenderProps, HasProps };
